requirements.txt
Django==4.0.5
django-extensions==3.1.5
djangorestframework==3.13.1
python-dotenv==0.20.0
